Decouple POSIX-specific code and provide graceful degradation.
