+++ CMakeLists.txt
@@ -89,6 +89,7 @@ if(RTLSDR_FOUND)
     list(APPEND SRCS src/runtime/worker_pool.cpp)
     list(APPEND SRCS src/runtime/config.cpp)
     list(APPEND SRCS src/runtime/log.cpp)
+    list(APPEND SRCS src/runtime/rt_sched.cpp)
     list(APPEND SRCS src/io/rtl_device.cpp)
     list(APPEND SRCS src/io/udp_control.cpp)
 endif()
