+++ include/runtime/rt_sched.h
@@ -1,0 +1,43 @@
+/*
+ * Realtime scheduling and CPU affinity helpers
+ * Copyright 2025 DSD-FME Authors
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef DSD_FME_RT_SCHED_H
+#define DSD_FME_RT_SCHED_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*
+ * Optionally enable realtime scheduling and set CPU affinity for the current
+ * thread based on environment variables.
+ *
+ * When `DSD_FME_RT_SCHED=1`, attempts to switch the calling thread to
+ * SCHED_FIFO with a priority derived from `DSD_FME_RT_PRIO_<ROLE>` if present.
+ * If `DSD_FME_CPU_<ROLE>` is set to a valid CPU index, pins the thread to that
+ * CPU.
+ *
+ * @param role Optional role label (e.g. "DEMOD", "DONGLE", "USB").
+ */
+void maybe_set_thread_realtime_and_affinity(const char* role);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* DSD_FME_RT_SCHED_H */
