+++ src/rtl_sdr_fm.cpp
@@ -22,7 +22,6 @@
  */
 
 #include <atomic>
-#include <errno.h>
 #include <math.h>
 #include <pthread.h>
 #include <rtl-sdr.h>
@@ -36,7 +35,6 @@
 #include "dsp/demod_pipeline.h"
 #include "dsp/fll.h"
 #include "dsp/resampler.h"
-#include "dsp/simd_widen.h"
 #include "dsp/ted.h"
 #include "io/rtl_device.h"
 #include "io/udp_control.h"
@@ -44,6 +42,7 @@
 #include "runtime/input_ring.h"
 #include "runtime/log.h"
 #include "runtime/ring.h"
+#include "runtime/rt_sched.h"
 #include "runtime/worker_pool.h"
 
 /* Runtime configuration documentation has moved to runtime/config.h. */
@@ -475,72 +474,7 @@ static struct input_ring_state input_ring;
  * @param role Optional role label (e.g. "DEMOD", "DONGLE") used to look up
  *             per-role environment variables.
  */
-static void
-maybe_set_thread_realtime_and_affinity(const char* role) {
-    const char* enable = getenv("DSD_FME_RT_SCHED");
-    if (!enable || enable[0] != '1') {
-        return;
-    }
-
-    /* Optional: role-specific priority (1..99) for SCHED_FIFO */
-    int policy = SCHED_FIFO;
-    struct sched_param sp;
-    int pmax = sched_get_priority_max(policy);
-    int pmin = sched_get_priority_min(policy);
-    int def = (pmax > 10) ? (pmax - 10) : pmax; /* default near top, but safe */
-    char envname[64];
-
-    sp.sched_priority = def;
-    if (role) {
-        /* e.g., DSD_FME_RT_PRIO_DEMOD, DSD_FME_RT_PRIO_DONGLE */
-        snprintf(envname, sizeof(envname), "DSD_FME_RT_PRIO_%s", role);
-        const char* prio_str = getenv(envname);
-        if (prio_str && prio_str[0] != '\0') {
-            int pr = atoi(prio_str);
-            if (pr < pmin) {
-                pr = pmin;
-            }
-            if (pr > pmax) {
-                pr = pmax;
-            }
-            sp.sched_priority = pr;
-        }
-    }
-
-    if (pthread_setschedparam(pthread_self(), policy, &sp) != 0) {
-        int err = errno;
-        LOG_WARNING("Failed to set %s thread to SCHED_FIFO (needs CAP_SYS_NICE). errno=%d (%s)\n", role ? role : "RT",
-                    err, strerror(err));
-    } else {
-        LOG_INFO("%s thread SCHED_FIFO priority set to %d.\n", role ? role : "RT", sp.sched_priority);
-    }
-
-    /* Optional: role-specific CPU affinity: DSD_FME_CPU_DEMOD / DSD_FME_CPU_DONGLE */
-    if (role) {
-        snprintf(envname, sizeof(envname), "DSD_FME_CPU_%s", role);
-        const char* cpu_str = getenv(envname);
-        if (cpu_str && cpu_str[0] != '\0') {
-            int cpu = atoi(cpu_str);
-            if (cpu >= 0) {
-#if defined(__linux__) && !defined(__CYGWIN__)
-                cpu_set_t cpuset;
-                CPU_ZERO(&cpuset);
-                CPU_SET((unsigned)cpu, &cpuset);
-                if (pthread_setaffinity_np(pthread_self(), sizeof(cpuset), &cpuset) != 0) {
-                    int err = errno;
-                    LOG_WARNING("Failed to set CPU affinity for %s thread to CPU %d. errno=%d (%s)\n", role, cpu, err,
-                                strerror(err));
-                } else {
-                    LOG_INFO("%s thread pinned to CPU %d.\n", role, cpu);
-                }
-#else
-                (void)cpu;
-                LOG_NOTICE("CPU affinity not supported on this platform.\n");
-#endif
-            }
-        }
-    }
-}
+/* moved to runtime/rt_sched.cpp */
 
 /* {length, coef, coef, coef}  and scaled by 2^15
    for now, only length 9, optimal way to get +85% bandwidth */
